Masih bug Itemnya

+++ app/src/main/java/com/project/pengelolakeuangan/data/TransactionData.kt
@@ -5,7 +5,7 @@ import java.time.LocalDate
 data class TransactionData(
     val id : Int,
     val isIncome: Boolean,
-    val nominal: Double,
+    val nominal: Long,
     val date: LocalDate,
     val method: String,
     val detail: String,
